% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{claims}
\alias{claims}
\title{Data frame with claims.
Taken with some data preparation from the Heritage Health Prize Competition.}
\description{
Data frame with claims.
Taken with some data preparation from the Heritage Health Prize Competition.
}
\keyword{data}
