% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{claims}
\alias{claims}
\title{Data frame with claims.
Taken with some data preparation from the Heritage Health Prize Competition.}
\description{
Data frame with claims.
Taken with some data preparation from the Heritage Health Prize Competition.
}
\keyword{data}

